+++ src/StubResult.php
@@ -2,7 +2,7 @@
 
 namespace DavidLienhard\Database;
 
-use DavidLienhard\Database\Exception as DatabaseException;
+use DavidLienhard\Database\Exceptions\Exception as DatabaseException;
 use DavidLienhard\Database\ResultInterface;
 use DavidLienhard\Database\ResultType;
 use DavidLienhard\Database\ResultTypeInterface;
